Removing comment code.
